Terminate linking trip to destination and secure content with user id

+++ app/model/DestinationModel.php
@@ -1,7 +1,31 @@
 <?php
     class DestinationModel extends Model {
 
+
+        private $idUser = '';
+
+        public function __construct(){
+            parent::__construct();
+            if(isset($_SESSION['idUser']))
+                $this->idUser = $_SESSION['idUser'];
+        }
+
+        private function isUserValid($idDestination = -1){
+            if($idDestination == -1){
+                $sth = $this->db->prepare('SELECT count(id) as num_rows FROM user WHERE id = :idUser');
+                $sth->execute([':idUser' => $this->idUser]);
+                return $sth->fetchAll()[0]['num_rows'] == 1;
+            }
+            else {
+                $sth = $this->db->prepare('SELECT count(destination.id) as num_rows FROM destination JOIN trip ON destination.id_trip = trip.id WHERE destination.id = :id AND id_user = :idUser');
+                $sth->execute([':id' => $idDestination, ':idUser' => $this->idUser]);
+                return $sth->fetchAll()[0]['num_rows'] == 1;
+            }
+        }
+
         public function getAllDestinations(){
+            if(!$this->isUserValid())
+                return [];
             return $this->db->query('SELECT * FROM destination')->fetchAll();
         }
 
@@ -10,17 +34,25 @@
         }
 
         public function getDestinationByID($id){
+            if(!$this->isUserValid($id))
+                return null;
             $sth = $this->db->prepare('SELECT * FROM destination WHERE id = :id');
             $sth->execute([':id' => $id]);
             return $sth->fetchAll();
         }
 
         public function deleteDestinationByID($id){
+            if(!$this->isUserValid($id))
+                return false;
             $sth = $this->db->prepare('DELETE FROM destination WHERE id = :id');
             $sth->execute([':id' => $id]);
+            return true;
         }
 
         public function addOrUpdate($destination){
+            if(!$this->isUserValid($destination['id']))
+                return false;
+
             if($destination['id'] == -1){
                 unset($destination['id']);
                 $sth = $this->db->prepare('INSERT INTO destination (name, lat, lng, description, startDate, endDate, id_transportation_type, id_trip)
